rename(components): 컴포넌트 네이밍 밎 구조 변경
- ui: html 기본 태그로만 구성된 기본 컴포넌트
- 그외
