+++ src/components/GoodsItem/Default/Default.stories.tsx
@@ -2,13 +2,13 @@ import type { Meta, StoryObj } from '@storybook/react';
 
 import { DEFAULT_IMAGE_URL } from '@/constants/data';
 
-import { RankingGoodsItem } from '@/components/ui/GoodsItem/Ranking';
+import { GoodsItem } from './Default';
 
-const meta: Meta<typeof RankingGoodsItem> = {
-  title: 'Example/GoodsItem/RankingGoodsItem',
+const meta: Meta<typeof GoodsItem> = {
+  title: 'Example/GoodsItem/Default',
   component: ({ ...props }) => (
     <div css={{ width: '8rem' }}>
-      <RankingGoodsItem {...props} />
+      <GoodsItem {...props} />
     </div>
   ),
   parameters: {
@@ -16,11 +16,6 @@ const meta: Meta<typeof RankingGoodsItem> = {
   },
   tags: ['autodocs'],
   argTypes: {
-    rank: {
-      control: 'number',
-      min: 1,
-      max: 100,
-    },
     imageSrc: {
       control: 'text',
     },
@@ -37,14 +32,13 @@ const meta: Meta<typeof RankingGoodsItem> = {
 };
 
 export default meta;
-type Story = StoryObj<typeof RankingGoodsItem>;
+type Story = StoryObj<typeof GoodsItem>;
 
 export const Default: Story = {
   args: {
-    rank: 1,
     imageSrc: DEFAULT_IMAGE_URL,
     subtitle: '카카오 프렌즈 특별 한정판',
     title: '[특가] 카카오 프렌즈 특별 한정판 브라이트 쿠션',
-    amount: '100000000000',
+    amount: 100000000000,
   },
 };
