+++ src/components/GoodsItem/Default/styles.ts
