+++ src/components/GoodsItem/Ranking/Ranking.stories.tsx
@@ -2,13 +2,13 @@ import type { Meta, StoryObj } from '@storybook/react';
 
 import { DEFAULT_IMAGE_URL } from '@/constants/data';
 
-import { GoodsItem } from '@/components/ui/GoodsItem/Default';
+import { RankingGoodsItem } from './Ranking';
 
-const meta: Meta<typeof GoodsItem> = {
-  title: 'Example/GoodsItem/Default',
+const meta: Meta<typeof RankingGoodsItem> = {
+  title: 'Example/GoodsItem/RankingGoodsItem',
   component: ({ ...props }) => (
     <div css={{ width: '8rem' }}>
-      <GoodsItem {...props} />
+      <RankingGoodsItem {...props} />
     </div>
   ),
   parameters: {
@@ -16,6 +16,11 @@ const meta: Meta<typeof GoodsItem> = {
   },
   tags: ['autodocs'],
   argTypes: {
+    rank: {
+      control: 'number',
+      min: 1,
+      max: 100,
+    },
     imageSrc: {
       control: 'text',
     },
@@ -32,13 +37,14 @@ const meta: Meta<typeof GoodsItem> = {
 };
 
 export default meta;
-type Story = StoryObj<typeof GoodsItem>;
+type Story = StoryObj<typeof RankingGoodsItem>;
 
 export const Default: Story = {
   args: {
+    rank: 1,
     imageSrc: DEFAULT_IMAGE_URL,
     subtitle: '카카오 프렌즈 특별 한정판',
     title: '[특가] 카카오 프렌즈 특별 한정판 브라이트 쿠션',
-    amount: '100000000000',
+    amount: 100000000000,
   },
 };
