+++ src/components/ui/Button/index.stories.ts
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react';
 
-import { Button } from '@/components/ui/Button';
+import { Button } from '.';
 
 const meta: Meta<typeof Button> = {
   title: 'Example/Button',
