+++ src/components/ui/Image/index.stories.ts
@@ -2,7 +2,7 @@ import type { Meta, StoryObj } from '@storybook/react';
 
 import { DEFAULT_IMAGE_URL } from '@/constants/data';
 
-import { Image } from '@/components/ui/Image/Default';
+import { Image } from './Default';
 
 const meta: Meta<typeof Image> = {
   title: 'Example/Image',
