+++ src/components/ui/Input/Input.stories.ts
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react';
 
-import { Input } from '@/components/ui/Input';
+import { Input } from '.';
 
 const meta: Meta<typeof Input> = {
   title: 'Example/Form/Input',
